+++ src/common/dto/token.response.ts
@@ -2,10 +2,12 @@ import { z } from 'zod';
 
 import { createResponseSchema } from '~/common/dto/response-schema';
 
+export const tokenSchema = z.object({
+    accessToken: z.string(),
+});
+
 export const TokenResponse = createResponseSchema(
-    z.object({
-        accessToken: z.string(),
-    }),
+    tokenSchema,
     'TokenResponseDto'
 );