+++ src/modules/auth/auth.repository.ts
@@ -1,8 +1,10 @@
 import { Injectable } from '@nestjs/common';
 
 import { PrismaService } from '~/infrastructure/database/prisma.service';
-import { AppRole } from '~/common/types/app-role.enum';
-import { userSelectAuth, UserWithPassword } from '~/common/types/auth.types';
+import {
+    userSelectAuth,
+    UserWithPassword,
+} from '~/common/types/auth.types';
 
 @Injectable()
 export class AuthRepository {
@@ -21,14 +23,6 @@ export class AuthRepository {
             return null;
         }
 
-        const owner = await this.prisma.owner.findUnique({
-            where: { userId: user.id },
-            select: { userId: true },
-        });
-
-        return {
-            ...user,
-            role: owner ? AppRole.OWNER : AppRole.USER,
-        };
+        return user;
     }
 }